+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Added `guidanceSourcesForPlans` query and `addPlanGuidance` and `removePlanGuidance` to `guidance` resolvers [#29]
 - Added a new `PlanGuidance` model and schema to accommodate the new `planGuidance` table [#29]
 - Added `acronyms` to `AffiliationSearch` schema so that client can get shortened names of affiliations [#29]
-- Added new methods to `guidanceServices`: `groupGuidanceByTag`, `getGuidanceSourcesForPlan`, `getSectionTags`,`getSectionTagsMap1` and `addPlanGuidanceAffiliation` [#29]
+- Added new methods to `guidanceServices`: `groupGuidanceByTag`, `getGuidanceSourcesForPlan`, `getSectionTags`, `getSectionTagIds`, `getSectionTagsMap`, `addPlanGuidanceAffiliation` and `getAffiliationsWithGuidanceForTemplate` [#29]
 - Added `ownerAffiliation` chained resolvers to `versionedQuestion` [#18]
 - added `@as-integrations/express5` for Apollo-Express integration